refactor(analytics): move analytics module to core package
