+++ app/src/main/java/com/geeksville/mesh/service/MeshService.kt
@@ -77,7 +77,7 @@ import kotlinx.coroutines.flow.flatMapLatest
 import kotlinx.coroutines.flow.flowOf
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
-import org.meshtastic.analytics.DataPair
+import org.meshtastic.core.analytics.DataPair
 import org.meshtastic.core.data.repository.MeshLogRepository
 import org.meshtastic.core.data.repository.NodeRepository
 import org.meshtastic.core.data.repository.PacketRepository
