+++ core/analytics/src/google/kotlin/org/meshtastic/core/analytics/di/GooglePlatformAnalyticsModule.kt
@@ -15,22 +15,21 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package org.meshtastic.analytics.di
+package org.meshtastic.core.analytics.di
 
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
-import org.meshtastic.analytics.platform.FdroidPlatformAnalytics
-import org.meshtastic.analytics.platform.PlatformAnalytics
+import org.meshtastic.core.analytics.platform.GooglePlatformAnalytics
+import org.meshtastic.core.analytics.platform.PlatformAnalytics
 import javax.inject.Singleton
 
-/** Hilt module to provide the [FdroidPlatformAnalytics] for the fdroid flavor. */
+/** Hilt module to provide the [GooglePlatformAnalytics] for the google flavor. */
 @Module
 @InstallIn(SingletonComponent::class)
-abstract class FdroidPlatformAnalyticsModule {
+abstract class GooglePlatformAnalyticsModule {
 
-    @Binds
-    @Singleton
-    abstract fun bindPlatformHelper(fdroidPlatformAnalytics: FdroidPlatformAnalytics): PlatformAnalytics
+    @Binds @Singleton
+    abstract fun bindPlatformHelper(googlePlatformHelper: GooglePlatformAnalytics): PlatformAnalytics
 }
