+++ core/analytics/src/google/kotlin/org/meshtastic/core/analytics/platform/GooglePlatformAnalytics.kt
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package org.meshtastic.analytics.platform
+package org.meshtastic.core.analytics.platform
 
 import android.app.Application
 import android.content.Context
@@ -59,8 +59,9 @@ import dagger.hilt.android.qualifiers.ApplicationContext
 import io.opentelemetry.api.GlobalOpenTelemetry
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
-import org.meshtastic.analytics.DataPair
+import org.meshtastic.core.analytics.DataPair
 import org.meshtastic.core.analytics.BuildConfig
+import org.meshtastic.core.analytics.platform.PlatformAnalytics
 import org.meshtastic.core.prefs.analytics.AnalyticsPrefs
 import timber.log.Timber
 import javax.inject.Inject
